+++ extractor/__init__.py
@@ -1,4 +1,0 @@
-from .cdm import *
-from .vendor import *
-
-__version__ = '1.1.0'